[ARCTIC-497] fix When adding cloumn to keyed hive table with partitions, the scheam sequence of change table and base table is different (#501)
* add HiveChangeInternalTable and BaseSchemaUpdate

* fix code style
