+++ hive/src/main/java/com/netease/arctic/hive/op/HiveSchemaUpdate.java
@@ -24,33 +24,24 @@ import com.netease.arctic.table.ArcticTable;
 import com.netease.arctic.table.KeyedTable;
 import com.netease.arctic.table.TableProperties;
 import org.apache.iceberg.FileFormat;
-import org.apache.iceberg.PartitionField;
-import org.apache.iceberg.Schema;
 import org.apache.iceberg.UpdateSchema;
-import org.apache.iceberg.types.Type;
-import org.apache.iceberg.types.Types;
 import org.apache.iceberg.util.PropertyUtil;
 
-import java.util.Collection;
 import java.util.Locale;
 
 /**
  * Schema evolution API implementation for {@link KeyedTable}.
  */
-public class HiveSchemaUpdate implements UpdateSchema {
+public class HiveSchemaUpdate extends BaseSchemaUpdate {
   private final ArcticTable arcticTable;
   private final HMSClientPool hiveClient;
   private final UpdateSchema updateSchema;
 
   public HiveSchemaUpdate(ArcticTable arcticTable, HMSClientPool hiveClient, UpdateSchema updateSchema) {
+    super(arcticTable, updateSchema);
     this.arcticTable = arcticTable;
     this.hiveClient = hiveClient;
     this.updateSchema = updateSchema;
-  }
-
-  @Override
-  public Schema apply() {
-    return this.updateSchema.apply();
   }
 
   @Override
@@ -71,112 +62,5 @@ public class HiveSchemaUpdate implements UpdateSchema {
         FileFormat.valueOf(PropertyUtil.propertyAsString(arcticTable.properties(), TableProperties.DEFAULT_FILE_FORMAT,
             TableProperties.DEFAULT_FILE_FORMAT_DEFAULT).toUpperCase(Locale.ENGLISH))));
     HiveTableUtil.persistTable(hiveClient, tbl);
-  }
-
-  @Override
-  public UpdateSchema allowIncompatibleChanges() {
-    throw new UnsupportedOperationException("hive table not support allowIncompatibleChanges");
-  }
-
-  @Override
-  public UpdateSchema addColumn(String name, Type type, String doc) {
-    this.updateSchema.addColumn(name, type, doc);
-    moveColBeforePar(name);
-    return this;
-  }
-
-  @Override
-  public UpdateSchema addColumn(String parent, String name, Type type, String doc) {
-    this.updateSchema.addColumn(parent, name, type, doc);
-    if (parent == null) {
-      moveColBeforePar(name);
-    }
-    return this;
-  }
-
-  @Override
-  public UpdateSchema addRequiredColumn(String name, Type type, String doc) {
-    throw new UnsupportedOperationException("hive table not support addRequiredColumn");
-  }
-
-  @Override
-  public UpdateSchema addRequiredColumn(String parent, String name, Type type, String doc) {
-    throw new UnsupportedOperationException("hive table not support addRequiredColumn");
-  }
-
-  @Override
-  public UpdateSchema renameColumn(String name, String newName) {
-    throw new UnsupportedOperationException("not support renameColumn now, there will be error when hive stored as " +
-        "parquet and we rename the column");
-  }
-
-  @Override
-  public UpdateSchema updateColumn(String name, Type.PrimitiveType newType) {
-    this.updateSchema.updateColumn(name, newType);
-    return this;
-  }
-
-  @Override
-  public UpdateSchema updateColumnDoc(String name, String newDoc) {
-    this.updateSchema.updateColumnDoc(name, newDoc);
-    return this;
-  }
-
-  @Override
-  public UpdateSchema makeColumnOptional(String name) {
-    throw new UnsupportedOperationException("hive table not support makeColumnOptional");
-  }
-
-  @Override
-  public UpdateSchema requireColumn(String name) {
-    throw new UnsupportedOperationException("hive table not support requireColumn");
-  }
-
-  @Override
-  public UpdateSchema deleteColumn(String name) {
-    throw new UnsupportedOperationException("hive table not support deleteColumn");
-  }
-
-  @Override
-  public UpdateSchema moveFirst(String name) {
-    throw new UnsupportedOperationException("hive table not support moveFirst");
-  }
-
-  @Override
-  public UpdateSchema moveBefore(String name, String beforeName) {
-    throw new UnsupportedOperationException("hive table not support moveBefore");
-  }
-
-  @Override
-  public UpdateSchema moveAfter(String name, String afterName) {
-    throw new UnsupportedOperationException("hive table not support moveAfter");
-  }
-
-  @Override
-  public UpdateSchema unionByNameWith(Schema newSchema) {
-    throw new UnsupportedOperationException("hive table not support unionByNameWith");
-  }
-
-  @Override
-  public UpdateSchema setIdentifierFields(Collection<String> names) {
-    throw new UnsupportedOperationException("hive table not support setIdentifierFields");
-  }
-
-  //It is strictly required that all non-partitioned columns precede partitioned columns in the schema.
-  private void moveColBeforePar(String name) {
-    if (!arcticTable.spec().isUnpartitioned()) {
-      int parFieldMinIndex = Integer.MAX_VALUE;
-      Types.NestedField firstParField = null;
-      for (PartitionField partitionField : arcticTable.spec().fields()) {
-        Types.NestedField sourceField = arcticTable.schema().findField(partitionField.sourceId());
-        if (arcticTable.schema().columns().indexOf(sourceField) < parFieldMinIndex) {
-          parFieldMinIndex = arcticTable.schema().columns().indexOf(sourceField);
-          firstParField = sourceField;
-        }
-      }
-      if (firstParField != null) {
-        this.updateSchema.moveBefore(name, firstParField.name());
-      }
-    }
   }
 }
