+++ hive/src/main/java/com/netease/arctic/hive/table/KeyedHiveTable.java
@@ -22,10 +22,16 @@ import com.netease.arctic.AmsClient;
 import com.netease.arctic.ams.api.TableMeta;
 import com.netease.arctic.hive.HMSClientPool;
 import com.netease.arctic.hive.HiveTableProperties;
+import com.netease.arctic.hive.op.BaseSchemaUpdate;
 import com.netease.arctic.hive.utils.HiveMetaSynchronizer;
+import com.netease.arctic.io.ArcticFileIO;
 import com.netease.arctic.table.BaseKeyedTable;
+import com.netease.arctic.table.BaseUnkeyedTable;
 import com.netease.arctic.table.ChangeTable;
 import com.netease.arctic.table.PrimaryKeySpec;
+import com.netease.arctic.table.TableIdentifier;
+import org.apache.iceberg.Table;
+import org.apache.iceberg.UpdateSchema;
 import org.apache.iceberg.util.PropertyUtil;
 
 /**
@@ -83,5 +89,19 @@ public class KeyedHiveTable extends BaseKeyedTable implements SupportHive {
   @Override
   public HMSClientPool getHMSClient() {
     return hiveClient;
+  }
+
+  public static class HiveChangeInternalTable extends BaseUnkeyedTable implements ChangeTable {
+
+    public HiveChangeInternalTable(
+        TableIdentifier tableIdentifier, Table changeIcebergTable, ArcticFileIO arcticFileIO,
+        AmsClient client) {
+      super(tableIdentifier, changeIcebergTable, arcticFileIO, client);
+    }
+
+    @Override
+    public UpdateSchema updateSchema() {
+      return new BaseSchemaUpdate(this, super.updateSchema());
+    }
   }
 }
